[MCUX-86685] [Edgefast] Fix type safety, input validation, and overflow protection in HPS service
This patch improves type safety in bit operations, adds input validation to
prevent buffer overflows, fixes data type sizing, and corrects code formatting
in hps.c and hps.h.

Key changes:
- Add HPS_BIT_SET() macro for type-safe bit setting with uint8_t cast
- Add HPS_BIT_CLEAR() macro for type-safe bit clearing with uint8_t cast
- write_http_headers(): Add buffer length validation when written by client (len > MAX_HEADERS_LEN)
- write_http_headers(): Return -ENOMEM when buffer size exceeded
- write_http_headers(): Replace bitwise operations with HPS_BIT_SET() for HEADERS_SET flag
- write_http_headers(): Replace bitwise operations with HPS_BIT_CLEAR() for HEADERS_SET flag
- write_http_headers(): Replace bitwise operations with HPS_BIT_SET() for HPS_HEADERS_RECEIVED
- write_http_headers(): Replace bitwise operations with HPS_BIT_SET() for HPS_HEADERS_TRUNCATED
- write_http_headers(): Replace bitwise operations with HPS_BIT_CLEAR() for HPS_HEADERS_TRUNCATED
- write_http_entity_body(): Add buffer length validation when written by client (len > MAX_BODY_LEN)
- write_http_entity_body(): Return -ENOMEM when buffer size exceeded
- write_http_entity_body(): Replace bitwise operations with HPS_BIT_SET() for BODY_SET flag
- write_http_entity_body(): Replace bitwise operations with HPS_BIT_CLEAR() for BODY_SET flag
- write_http_entity_body(): Replace bitwise operations with HPS_BIT_SET() for HPS_BODY_RECEIVED
- write_http_entity_body(): Replace bitwise operations with HPS_BIT_SET() for HPS_BODY_TRUNCATED
- write_http_entity_body(): Replace bitwise operations with HPS_BIT_CLEAR() for HPS_BODY_TRUNCATED
- write_uri(): Add buffer length validation (len > MAX_URI_LEN)
- write_uri(): Return -BT_ATT_ERR_INVALID_ATTRIBUTE_LEN when URI length exceeded
- write_uri(): Replace bitwise operations with HPS_BIT_SET() for URI_SET flag
- bt_hps_set_status_code(): Replace bitwise operations with HPS_BIT_CLEAR() for URI_SET flag
- hps_config_t: Change uri_len type from uint8_t to uint16_t to match MAX_URI_LEN size
- Fix trailing whitespace in HTTP_REQ_CANCEL enum
- Fix trailing whitespace in hps_flags_t field

These changes prevent buffer overflows by validating input lengths before
copying, improve type safety by using explicit casts in bit operation macros,
fix potential truncation by using uint16_t for URI length, and ensure
consistent code formatting.

+++ include/bluetooth/services/hps.h
@@ -42,7 +42,7 @@ enum
     HTTPS_POST_REQ,
     HTTPS_PUT_REQ,
     HTTPS_DELETE_REQ,
-    HTTP_REQ_CANCEL,  
+    HTTP_REQ_CANCEL,
 };
 
 enum
@@ -84,7 +84,7 @@ typedef struct
 
 typedef struct
 {
-    uint8_t                 uri_len;
+    uint16_t                uri_len;
     uint8_t                 uri[MAX_URI_LEN];
     uint16_t                http_headers_len;
     uint8_t                 http_headers[MAX_HEADERS_LEN];
@@ -93,7 +93,7 @@ typedef struct
     hps_status_t            status_code;
     hps_http_command_t      command;
     hps_state_t             state;
-    hps_flags_t             flags;    
+    hps_flags_t             flags;
 } hps_config_t;
 
 #ifdef __cplusplus
